requirements.txt
mypy
python-docx
nltk
tika
spacy
numpy
pandas
gensim
sklearn
scipy
